base config for defaultconfig, cleanups

+++ internal/cli/defaultconfig.go
@@ -18,30 +18,36 @@ import (
 
 func DefaultConfigCommand() *cobra.Command {
 	var defaultConfigFile string
+	var baseConfigFile string
+	var dryRun bool
 	var defaultConfigCmd = &cobra.Command{
 		Use:   "defaultconfig",
 		Short: "Generate full configuration file with defaults",
 		Long:  `Generate full Centrifugo configuration file with defaults`,
 		Run: func(cmd *cobra.Command, args []string) {
-			DefaultConfig(defaultConfigFile)
+			DefaultConfig(defaultConfigFile, baseConfigFile, dryRun)
 		},
 	}
 	defaultConfigCmd.Flags().StringVarP(&defaultConfigFile, "config", "c", "config.json", "path to default config file to generate")
+	defaultConfigCmd.Flags().StringVarP(&baseConfigFile, "base", "b", "", "path to the base config file to use")
+	defaultConfigCmd.Flags().BoolVarP(&dryRun, "dry-run", "d", false, "dry run mode (do not write file and just print generated config to stdout)")
 	return defaultConfigCmd
 }
 
-func DefaultConfig(configFile string) {
-	exists, err := tools.PathExists(configFile)
-	if err != nil {
-		fmt.Printf("error: %v\n", err)
-		os.Exit(1)
+func DefaultConfig(configFile string, baseFile string, dryRun bool) {
+	if !dryRun {
+		exists, err := tools.PathExists(configFile)
+		if err != nil {
+			fmt.Printf("error: %v\n", err)
+			os.Exit(1)
+		}
+		if exists {
+			fmt.Printf("error: target file already exists\n")
+			os.Exit(1)
+		}
 	}
-	if exists {
-		fmt.Printf("error: target file already exists\n")
-		os.Exit(1)
-	}
-	conf, _, _ := config.GetConfig(nil, "")
-	if err = conf.Validate(); err != nil {
+	conf, _, _ := config.GetConfig(nil, baseFile)
+	if err := conf.Validate(); err != nil {
 		fmt.Printf("error: %v\n", err)
 		os.Exit(1)
 	}
@@ -61,13 +67,6 @@ func DefaultConfig(configFile string) {
 		os.Exit(1)
 	}
 
-	jsonMap := make(map[string]interface{})
-	err = json.Unmarshal(jsonBytes, &jsonMap)
-	if err != nil {
-		fmt.Printf("error: %v\n", err)
-		os.Exit(1)
-	}
-
 	switch ext {
 	case "json":
 		b = jsonBytes
@@ -82,6 +81,11 @@ func DefaultConfig(configFile string) {
 	if err != nil {
 		fmt.Printf("error: %v\n", err)
 		os.Exit(1)
+	}
+
+	if dryRun {
+		fmt.Println(string(b))
+		return
 	}
 
 	err = os.WriteFile(configFile, b, 0644)
